requirements.txt
fastapi
uvicorn
pandas
numpy
xgboost
keras
tensorflow
stable-baselines3
gymnasium
selenium
beautifulsoup4
sentence-transformers
requests
fuzzywuzzy
googlesearch-python
webdriver-manager
pytest
